+++ src/assets/nowplaying_data.js
@@ -1,0 +1,74 @@
+export const NowPlaying_Data = [
+  {
+    "game_id": "rivals",
+    "game_title": "Marvel Rivals",
+    "stream_title": "Pure Chaos!!",
+    "schedule": null,
+    "description": "Superheroes, supervillains, and super chaos! Watch as @the13thgeek assembles the ultimate team... or gets yeeted across the battlefield.",
+    "show_game": true
+  },
+  {
+    "game_id": "haloce",
+    "game_title": "Halo: Combat Evolved",
+    "stream_title": "Silent Cartographer",
+    "schedule": "Tuesdays @ 7pm EST",
+    "description": "The Master Chief's first mission begins! @the13thgeek takes on the Covenent, but will he save the galaxy or get lost in space?",
+    "show_game": true
+  },
+  {
+    "game_id": "tpmuseum",
+    "game_title": "Two Point Museum",
+    "stream_title": "Honey, Where's My Donation Box?",
+    "schedule": "Thursdays @ 7pm EST",
+    "description": "Building a world-class museum sounds easy... until @the13thgeek gets involved. Expect exhibits, chaos, and questionable curation!",
+    "show_game": true
+  },
+  {
+    "game_id": "ritarewind",
+    "game_title": "Mighty Morphin Power Rangers: Rita's Rewind",
+    "stream_title": "It's Morphin' Time!",
+    "schedule": null,
+    "description": "Bring out your Morphers and join @the13thgeek and friends as they morph into action and retake Angel Grove!",
+    "show_game": false
+  },
+  {
+    "game_id": "smpjamboree",
+    "game_title": "Super Mario Party Jamboree",
+    "stream_title": "The Party Is Back!",
+    "schedule": null,
+    "description": "Come for the stars, stay for the mayhem. Friendships will be put to the ultimate test... again.",
+    "show_game": false
+  },
+  {
+    "game_id": "b&b",
+    "game_title": "Bear & Breakfast",
+    "stream_title": "AirGeek B&B",
+    "schedule": null,
+    "description": "@the13thgeek's running a cozy inn... as a teddy bear! Watch as he balances hospitality, chaos and fur-covered pancakes!",
+    "show_game": false
+  },
+  {
+    "game_id": "pc2",
+    "game_title": "Planet Coaster 2",
+    "stream_title": "Splash'n Fun",
+    "schedule": null,
+    "description": "Let's check out the new Planet Coaster sequel - now with water parks! Don't forget to bring sunblock.",
+    "show_game": false
+  },
+  {
+    "game_id": "re5",
+    "game_title": "Resident Evil 5",
+    "stream_title": "Kijuju Carnage",
+    "schedule": null,
+    "description": "@the13thgeek and friends face zombies, explosions, and co-op madness. Will they survive? Probably not, but it'll be epic!",
+    "show_game": false
+  },
+  {
+    "game_id": "helldivers2",
+    "game_title": "Helldivers II",
+    "stream_title": "Say Hello to Democracy!",
+    "schedule": null,
+    "description": "@the13thgeek teams up with streamers to take on alien hordes in space, causing chaos and accidental explosions. Friendly fire? Definitely.",
+    "show_game": true
+  } 
+];